keep me signed in logic

+++ frontend-admin/app/composables/useAuth.js
@@ -7,8 +7,11 @@ export const useAuth = () => {
 
     const loadUser = () => {
         if (process.client) {
-            const raw = localStorage.getItem('user')
-            user.value = raw ? JSON.parse(raw) : null
+            const local_raw = localStorage.getItem('user')
+            if (!local_raw) {
+                const session_raw = sessionStorage.getItem('user')
+                user.value = session_raw ? JSON.parse(session_raw) : null
+            } else user.value = local_raw ? JSON.parse(local_raw) : null
         }
     }
 
@@ -19,22 +22,31 @@ export const useAuth = () => {
             body: { phone_number, password, keep_me_signed_in },
         })
         if (res.access && res.user) {
-            localStorage.setItem('access_token', res.access)
-            localStorage.setItem('refresh_token', res.refresh)
-            localStorage.setItem('user', JSON.stringify(res.user))
+            if (keep_me_signed_in) {
+                localStorage.setItem('access_token', res.access)
+                localStorage.setItem('refresh_token', res.refresh)
+                localStorage.setItem('user', JSON.stringify(res.user))
+            } else {
+                sessionStorage.setItem('access_token', res.access)
+                sessionStorage.setItem('refresh_token', res.refresh)
+                sessionStorage.setItem('user', JSON.stringify(res.user))
+            }
             user.value = res.user
             await router.push('/dashboard')
         }
     }
 
     const logout = async () => {
-        const refresh = process.client ? localStorage.getItem('refresh_token') : null
+        const local_refresh = process.client ? localStorage.getItem('refresh_token') : null
+        let session_refresh
+        if (!local_refresh) session_refresh = process.client ? sessionStorage.getItem('refresh_token') : null
+
         const config = useRuntimeConfig()
-        if (refresh) {
+        if (local_refresh || session_refresh) {
             try {
                 await $fetch(`${config.public.apiBase}user/sign-out-request/`, {
                     method: 'POST',
-                    body: { refresh },
+                    body: { refresh: local_refresh ? local_refresh : session_refresh },
                     headers: {
                         Authorization: `Bearer ${getAccessToken()}`,
                     },
@@ -44,25 +56,35 @@ export const useAuth = () => {
             }
         }
         user.value = null
-        if (process.client) localStorage.clear()
+        if (process.client){
+            localStorage.clear()
+            sessionStorage.clear()
+        }
         await router.push('/login')
     }
 
-    const getAccessToken = () => (process.client ? localStorage.getItem('access_token') : null)
+    const getAccessToken = () => {
+        let token = process.client ? localStorage.getItem('access_token') : null
+        if (!token) token = process.client ? sessionStorage.getItem('access_token') : null
+        return token
+    }
 
     const refreshToken = async () => {
         if (!process.client) return false
-        const token = localStorage.getItem('refresh_token')
-        if (!token) return false
-
+        let local_token, session_token
+        local_token = localStorage.getItem('refresh_token')
+        if (!local_token) {
+            session_token = sessionStorage.getItem('refresh_token')
+            if (!session_token) return false
+        }
         const config = useRuntimeConfig()
         try {
             const res = await $fetch(`${config.public.apiBase}token/refresh/`, {
                 method: 'POST',
-                body: { refresh: token },
+                body: { refresh: local_token ? local_token : session_token },
             })
             if (res.access) {
-                localStorage.setItem('access_token', res.access)
+                local_token ? localStorage.setItem('access_token', res.access) : sessionStorage.setItem('access_token', res.access)
                 return true
             }
             return false
